refactor(docker): split driver into cohesive files and tighten signatures
Split pkg/driver/docker/docker.go (1322 LOC) into responsibility-grouped
files matching the package's existing build.go/diagnostics.go convention:
- docker.go: struct, factory, core helpers
- lifecycle.go: container lifecycle operations
- runargs.go: docker run argument construction
- useruid.go: container UID/GID remapping

Idiomatic cleanups (no new abstractions):
- Flatten runArgsBuilder's half-fluent chain into a uniform step list.
- Collapse the per-IDE JetBrains switch into a constructor lookup table.
- Rename the config2/provider2 import aliases to pkgconfig/provider.

Promote CommandDevContainer's 7 positional args to a driver.CommandParams
struct on the Driver interface, consistent with RunDockerDevContainerParams.
Updates the docker, custom, and kubernetes drivers, the delivery adapter,
and all call sites.

Tests split to mirror the source (runargs_test.go, useruid_test.go).
