+++ pkg/driver/docker/docker.go
@@ -1,24 +1,23 @@
 package docker
 
 import (
-	"bufio"
 	"context"
 	"fmt"
-	"io"
-	"os"
-	"os/user"
 	"runtime"
-	"strconv"
-	"strings"
 
 	"github.com/devsy-org/devsy/pkg/compose"
-	config2 "github.com/devsy-org/devsy/pkg/config"
 	"github.com/devsy-org/devsy/pkg/devcontainer/config"
 	"github.com/devsy-org/devsy/pkg/docker"
 	"github.com/devsy-org/devsy/pkg/driver"
-	"github.com/devsy-org/devsy/pkg/ide/jetbrains"
 	"github.com/devsy-org/devsy/pkg/log"
-	provider2 "github.com/devsy-org/devsy/pkg/provider"
+	"github.com/devsy-org/devsy/pkg/provider"
+)
+
+const (
+	// dockerExec is the docker subcommand for running a command in a container.
+	dockerExec = "exec"
+	// rootUser is the conventional root account name/uid-0 owner.
+	rootUser = "root"
 )
 
 func makeEnvironment(env map[string]string) []string {
@@ -35,7 +34,7 @@ func makeEnvironment(env map[string]string) []string {
 }
 
 func NewDockerDriver(
-	workspaceInfo *provider2.AgentWorkspaceInfo,
+	workspaceInfo *provider.AgentWorkspaceInfo,
 ) (driver.DockerDriver, error) {
 	dockerCommand := "docker"
 	if workspaceInfo.Agent.Docker.Path != "" {
@@ -83,151 +82,6 @@ type dockerDriver struct {
 
 func (d *dockerDriver) TargetArchitecture(ctx context.Context, workspaceId string) (string, error) {
 	return runtime.GOARCH, nil
-}
-
-func (d *dockerDriver) CommandDevContainer(
-	ctx context.Context,
-	workspaceId, user, command string,
-	stdin io.Reader,
-	stdout io.Writer,
-	stderr io.Writer,
-) error {
-	container, err := d.FindDevContainer(ctx, workspaceId)
-	if err != nil {
-		return err
-	} else if container == nil {
-		return fmt.Errorf("container not found")
-	}
-
-	status := strings.ToLower(container.State.Status)
-	if status == "dead" || status == "removing" {
-		return fmt.Errorf(
-			"%w: container %s is %q",
-			docker.ErrContainerTerminal,
-			container.ID,
-			status,
-		)
-	}
-	if status != "running" {
-		log.Infof(
-			"container %s is not running (status=%s), restarting",
-			container.ID, status,
-		)
-		if err := d.Docker.StartContainer(ctx, container.ID); err != nil {
-			return fmt.Errorf("restart container: %w", err)
-		}
-		if err := d.Docker.WaitContainerRunning(ctx, container.ID); err != nil {
-			return fmt.Errorf("wait for container to be running: %w", err)
-		}
-		log.Infof("container %s is now running", container.ID)
-	}
-
-	args := []string{"exec"}
-	if stdin != nil {
-		args = append(args, "-i")
-	}
-	args = append(args, "-u", user, container.ID, "sh", "-c", command)
-	return d.Docker.Run(ctx, args, stdin, stdout, stderr)
-}
-
-func (d *dockerDriver) PushDevContainer(ctx context.Context, image string) error {
-	// push image
-	writer := log.Writer(log.LevelInfo)
-	defer func() { _ = writer.Close() }()
-
-	// build args
-	args := []string{
-		"push",
-		image,
-	}
-
-	// run command
-	log.Debugf(
-		"running docker push command: command=%s, args=%s",
-		d.Docker.DockerCommand,
-		strings.Join(args, " "),
-	)
-	err := d.Docker.Run(ctx, args, nil, writer, writer)
-	if err != nil {
-		return fmt.Errorf("push image: %w", err)
-	}
-
-	return nil
-}
-
-func (d *dockerDriver) TagDevContainer(ctx context.Context, image, tag string) error {
-	// Tag image
-	writer := log.Writer(log.LevelInfo)
-	defer func() { _ = writer.Close() }()
-
-	// build args
-	args := []string{
-		"tag",
-		image,
-		tag,
-	}
-
-	// run command
-	log.Debugf(
-		"running docker tag command: command=%s, args=%s",
-		d.Docker.DockerCommand,
-		strings.Join(args, " "),
-	)
-	err := d.Docker.Run(ctx, args, nil, writer, writer)
-	if err != nil {
-		return fmt.Errorf("tag image: %w", err)
-	}
-
-	return nil
-}
-
-func (d *dockerDriver) DeleteDevContainer(ctx context.Context, workspaceId string) error {
-	container, err := d.FindDevContainer(ctx, workspaceId)
-	if err != nil {
-		return err
-	} else if container == nil {
-		return nil
-	}
-
-	err = d.Docker.Remove(ctx, container.ID)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (d *dockerDriver) StartDevContainer(ctx context.Context, workspaceId string) error {
-	container, err := d.FindDevContainer(ctx, workspaceId)
-	if err != nil {
-		return err
-	} else if container == nil {
-		return fmt.Errorf("container not found")
-	}
-
-	return d.Docker.StartContainer(ctx, container.ID)
-}
-
-func (d *dockerDriver) StopDevContainer(ctx context.Context, workspaceId string) error {
-	container, err := d.FindDevContainer(ctx, workspaceId)
-	if err != nil {
-		return err
-	} else if container == nil {
-		return fmt.Errorf("container not found")
-	}
-
-	return d.Docker.Stop(ctx, container.ID)
-}
-
-func (d *dockerDriver) InspectImage(
-	ctx context.Context,
-	imageName string,
-) (*config.ImageDetails, error) {
-	return d.Docker.InspectImage(ctx, imageName, true)
-}
-
-func (d *dockerDriver) GetImageTag(ctx context.Context, imageID string) (string, error) {
-	return d.Docker.GetImageTag(ctx, imageID)
 }
 
 func (d *dockerDriver) ComposeHelper() (*compose.ComposeHelper, error) {
@@ -278,1045 +132,4 @@ func (d *dockerDriver) FindDevContainer(
 	}
 
 	return containerDetails, nil
-}
-
-func (d *dockerDriver) RunDevContainer(
-	ctx context.Context,
-	workspaceId string,
-	options *driver.RunOptions,
-) error {
-	return fmt.Errorf("unsupported")
-}
-
-func (d *dockerDriver) RunDockerDevContainer(
-	ctx context.Context,
-	params *driver.RunDockerDevContainerParams,
-) error {
-	if err := d.EnsureImage(ctx, params.Options); err != nil {
-		return err
-	}
-
-	helper, err := d.DockerHelper()
-	if err != nil {
-		return err
-	}
-
-	args, err := d.buildRunArgs(params, helper)
-	if err != nil {
-		return err
-	}
-
-	writer := log.Writer(log.LevelInfo)
-	defer func() { _ = writer.Close() }()
-
-	if err := d.startContainer(ctx, params.LocalWorkspaceFolder, args, writer); err != nil {
-		return err
-	}
-
-	return d.UpdateContainerUserUID(ctx, params.WorkspaceID, params.ParsedConfig, writer)
-}
-
-func (d *dockerDriver) EnsureImage(
-	ctx context.Context,
-	options *driver.RunOptions,
-) error {
-	log.Infof("inspecting image: image=%s", options.Image)
-	_, err := d.Docker.InspectImage(ctx, options.Image, false)
-	if err != nil {
-		log.Infof("image not found, pulling image: image=%s", options.Image)
-		writer := log.Writer(log.LevelDebug)
-		defer func() { _ = writer.Close() }()
-
-		return d.Docker.Pull(ctx, docker.PullOptions{
-			Image:    options.Image,
-			Platform: options.Platform,
-			Stdout:   writer,
-			Stderr:   writer,
-		})
-	}
-	return nil
-}
-
-func (d *dockerDriver) EnsurePath(path *config.Mount) *config.Mount {
-	// in case of local windows and remote linux tcp, we need to manually do the path conversion
-	if runtime.GOOS == "windows" {
-		for _, v := range d.Docker.Environment {
-			// we do this only is DOCKER_HOST is not docker-desktop engine, but
-			// a direct TCP connection to a docker daemon running in WSL
-			if strings.Contains(v, "DOCKER_HOST=tcp://") {
-				unixPath := path.Source
-				unixPath = strings.Replace(unixPath, "C:", "c", 1)
-				unixPath = strings.ReplaceAll(unixPath, "\\", "/")
-				unixPath = "/mnt/" + unixPath
-
-				path.Source = unixPath
-
-				return path
-			}
-		}
-	}
-	return path
-}
-
-func (d *dockerDriver) GetDevContainerLogs(
-	ctx context.Context,
-	workspaceId string,
-	stdout io.Writer,
-	stderr io.Writer,
-) error {
-	container, err := d.FindDevContainer(ctx, workspaceId)
-	if err != nil {
-		return err
-	} else if container == nil {
-		return fmt.Errorf("container not found")
-	}
-
-	return d.Docker.GetContainerLogs(ctx, container.ID, stdout, stderr)
-}
-
-func (d *dockerDriver) UpdateContainerUserUID(
-	ctx context.Context,
-	workspaceId string,
-	parsedConfig *config.DevContainerConfig,
-	writer io.Writer,
-) error {
-	if !d.shouldUpdateUserUID(parsedConfig) {
-		return nil
-	}
-
-	localUser, containerUser, err := d.gatherUpdateRequirements(parsedConfig)
-	if err != nil {
-		return err
-	}
-	// containerUser is guaranteed non-empty by shouldUpdateUserUID
-
-	if localUser.Uid == "0" {
-		log.Info("local user is root, skipping UID/GID update")
-		return nil
-	}
-
-	container, err := d.FindDevContainer(ctx, workspaceId)
-	if err != nil {
-		return err
-	}
-	if container == nil {
-		return fmt.Errorf("container not found")
-	}
-
-	files, info, err := d.updateUserMappings(ctx, &userMappingParams{
-		containerID:   container.ID,
-		containerUser: containerUser,
-		localUser:     localUser,
-		writer:        writer,
-	})
-	if err != nil {
-		return err
-	}
-	defer files.cleanup()
-
-	if shouldSkipUpdate(localUser, info) {
-		log.Info("container user UID/GID already match local user, skipping update")
-		return nil
-	}
-
-	log.Infof("updating container user %q UID from %s to %s and GID from %s to %s",
-		containerUser, info.Uid, localUser.Uid, info.Gid, localUser.Gid)
-
-	if err := d.copyFilesToContainer(ctx, container.ID, files, writer); err != nil {
-		return err
-	}
-
-	return d.applyPermissions(ctx, container.ID, localUser.Uid, localUser.Gid, info.HomeDir, writer)
-}
-
-func (d *dockerDriver) gatherUpdateRequirements(
-	parsedConfig *config.DevContainerConfig,
-) (*user.User, string, error) {
-	localUser, err := user.Current()
-	if err != nil {
-		return nil, "", err
-	}
-
-	containerUser := d.getContainerUser(parsedConfig)
-	return localUser, containerUser, nil
-}
-
-type userMappingParams struct {
-	containerID   string
-	containerUser string
-	localUser     *user.User
-	writer        io.Writer
-}
-
-func (d *dockerDriver) updateUserMappings(
-	ctx context.Context,
-	params *userMappingParams,
-) (*tempFiles, *user.User, error) {
-	files, err := d.createTempFiles()
-	if err != nil {
-		return nil, nil, err
-	}
-
-	if err := d.fetchContainerFiles(ctx, params.containerID, files, params.writer); err != nil {
-		files.cleanup()
-		return nil, nil, err
-	}
-
-	info, err := d.processUserFiles(
-		files,
-		params.containerUser,
-		params.localUser.Uid,
-		params.localUser.Gid,
-	)
-	if err != nil {
-		files.cleanup()
-		return nil, nil, err
-	}
-
-	return files, info, nil
-}
-
-// shouldSkipUpdate returns true if UID/GID mapping should be skipped.
-// localUser is the host system's current user.
-// info contains the container user's current UID/GID (parsed from container's /etc/passwd).
-func shouldSkipUpdate(localUser *user.User, info *user.User) bool {
-	return info.Uid == "0" || (localUser.Uid == info.Uid && localUser.Gid == info.Gid)
-}
-
-type runArgsBuilder struct {
-	args   []string
-	driver *dockerDriver
-	params *driver.RunDockerDevContainerParams
-}
-
-func (d *dockerDriver) buildRunArgs(
-	params *driver.RunDockerDevContainerParams,
-	helper *docker.DockerHelper,
-) ([]string, error) {
-	b := &runArgsBuilder{
-		args:   []string{"run"},
-		driver: d,
-		params: params,
-	}
-
-	if helper.GetRuntime().SupportsSignalProxy() {
-		b.args = append(b.args, "--sig-proxy=false")
-	}
-
-	b.addPorts().
-		addWorkspaceMount(helper).
-		addUser().
-		addEnv().
-		addInit().
-		addPrivileged()
-
-	if err := b.addPodmanArgs(); err != nil {
-		return nil, err
-	}
-
-	b.addCapabilities()
-
-	if err := b.addMounts(); err != nil {
-		return nil, err
-	}
-
-	b.addIDEMount().
-		addLabels().
-		addGPU().
-		addRunArgs().
-		addRunPlatform().
-		addDetached().
-		addEntrypoint().
-		addImage()
-
-	return b.args, nil
-}
-
-func (b *runArgsBuilder) addPorts() *runArgsBuilder {
-	b.args = b.driver.addPortArgs(b.args, b.params.ParsedConfig)
-	return b
-}
-
-func (b *runArgsBuilder) addWorkspaceMount(helper *docker.DockerHelper) *runArgsBuilder {
-	b.args = b.driver.addWorkspaceMountArgs(b.args, b.params.Options, helper)
-	return b
-}
-
-func (b *runArgsBuilder) addUser() *runArgsBuilder {
-	b.args = b.driver.addUserArgs(b.args, b.params.Options)
-	return b
-}
-
-func (b *runArgsBuilder) addEnv() *runArgsBuilder {
-	b.args = b.driver.addEnvArgs(b.args, b.params.Options)
-	return b
-}
-
-func (b *runArgsBuilder) addInit() *runArgsBuilder {
-	b.args = b.driver.addInitArgs(b.args, b.params.Options)
-	return b
-}
-
-func (b *runArgsBuilder) addPrivileged() *runArgsBuilder {
-	b.args = b.driver.addPrivilegedArgs(b.args, b.params.Options)
-	return b
-}
-
-func (b *runArgsBuilder) addPodmanArgs() error {
-	podmanArgs, err := b.driver.getPodmanArgs(b.params.Options, b.params.ParsedConfig)
-	if err != nil {
-		return err
-	}
-	b.args = append(b.args, podmanArgs...)
-	return nil
-}
-
-func (b *runArgsBuilder) addCapabilities() *runArgsBuilder {
-	b.args = b.driver.addCapabilityArgs(b.args, b.params.Options)
-	return b
-}
-
-func (b *runArgsBuilder) addMounts() error {
-	args, err := b.driver.addMountArgs(b.args, b.params.Options)
-	if err != nil {
-		return err
-	}
-	b.args = args
-	return nil
-}
-
-func (b *runArgsBuilder) addIDEMount() *runArgsBuilder {
-	b.args = b.driver.addIDEMountArgs(b.args, b.params.IDE, b.params.IDEOptions)
-	return b
-}
-
-func (b *runArgsBuilder) addLabels() *runArgsBuilder {
-	b.args = b.driver.addLabelArgs(b.args, b.params.WorkspaceID, b.params.Options)
-	return b
-}
-
-func (b *runArgsBuilder) addGPU() *runArgsBuilder {
-	b.args = appendGPUOptions(b.params.ParsedConfig, b.driver, b.args, b.params.GPUAvailability)
-	return b
-}
-
-func (b *runArgsBuilder) addRunArgs() *runArgsBuilder {
-	b.args = append(b.args, b.params.ParsedConfig.RunArgs...)
-	return b
-}
-
-func (b *runArgsBuilder) addRunPlatform() *runArgsBuilder {
-	if b.params.Options.Platform == "" {
-		return b
-	}
-	for _, a := range b.params.ParsedConfig.RunArgs {
-		if a == "--platform" || strings.HasPrefix(a, "--platform=") {
-			return b // explicit config wins; avoid duplicate/conflict.
-		}
-	}
-	b.args = append(b.args, "--platform="+b.params.Options.Platform)
-	return b
-}
-
-func (b *runArgsBuilder) addDetached() *runArgsBuilder {
-	b.args = append(b.args, "-d")
-	return b
-}
-
-func (b *runArgsBuilder) addEntrypoint() *runArgsBuilder {
-	b.args = b.driver.addEntrypointArgs(b.args, b.params.Options)
-	return b
-}
-
-func (b *runArgsBuilder) addImage() *runArgsBuilder {
-	b.args = append(b.args, b.params.Options.Image)
-	b.args = append(b.args, b.params.Options.Cmd...)
-	return b
-}
-
-func (d *dockerDriver) addPortArgs(
-	args []string,
-	parsedConfig *config.DevContainerConfig,
-) []string {
-	for _, appPort := range parsedConfig.AppPort {
-		intPort, err := strconv.Atoi(appPort)
-		if err != nil {
-			args = append(args, "-p", appPort)
-		} else {
-			args = append(args, "-p", fmt.Sprintf("127.0.0.1:%d:%d", intPort, intPort))
-		}
-	}
-	return args
-}
-
-func (d *dockerDriver) addWorkspaceMountArgs(
-	args []string,
-	options *driver.RunOptions,
-	helper *docker.DockerHelper,
-) []string {
-	if options.WorkspaceMount != nil {
-		workspacePath := d.EnsurePath(options.WorkspaceMount)
-		mountPath := workspacePath.String()
-		if !helper.GetRuntime().SupportsMountConsistency() {
-			mountPath = stripMountConsistency(mountPath)
-		}
-		if shouldBustBindCache(helper) {
-			mountPath = withBindCreateSrc(mountPath)
-		}
-		args = append(args, "--mount", mountPath)
-	}
-	return args
-}
-
-// minBindCreateSrcMajor is the docker CLI major version that added the
-// bind-create-src mount option; older clients reject it at parse time.
-const minBindCreateSrcMajor = 29
-
-// shouldBustBindCache reports whether to add bind-create-src to the workspace
-// mount. Gated to Docker (Podman/nerdctl reject it), Docker Desktop
-// (GOOS != linux), and CLI >= v29.
-func shouldBustBindCache(helper *docker.DockerHelper) bool {
-	if helper.GetRuntime().Name() != docker.RuntimeDocker {
-		return false
-	}
-	if runtime.GOOS == "linux" {
-		return false
-	}
-	return dockerMajorAtLeast(helper.ClientVersion(context.Background()), minBindCreateSrcMajor)
-}
-
-// dockerMajorAtLeast reports whether version's major component is >= minMajor.
-// An unparseable version returns false.
-func dockerMajorAtLeast(version string, minMajor int) bool {
-	major, _, ok := strings.Cut(version, ".")
-	if !ok {
-		return false
-	}
-	n, err := strconv.Atoi(strings.TrimSpace(major))
-	if err != nil {
-		return false
-	}
-	return n >= minMajor
-}
-
-// withBindCreateSrc adds bind-create-src=true to a bind --mount whose source
-// exists, forcing Docker Desktop to re-resolve a stale-cached path. A missing
-// source is left untouched so docker still fails loudly rather than binding an
-// empty dir.
-func withBindCreateSrc(mountPath string) string {
-	m := config.ParseMount(mountPath)
-	if m.Type != "bind" || m.Source == "" {
-		return mountPath
-	}
-	for part := range strings.SplitSeq(mountPath, ",") {
-		if strings.HasPrefix(part, "bind-create-src=") {
-			return mountPath
-		}
-	}
-	if _, err := os.Stat(m.Source); err != nil {
-		return mountPath
-	}
-	return mountPath + ",bind-create-src=true"
-}
-
-func stripMountConsistency(mount string) string {
-	var parts []string
-	for part := range strings.SplitSeq(mount, ",") {
-		if !strings.HasPrefix(part, "consistency=") {
-			parts = append(parts, part)
-		}
-	}
-	return strings.Join(parts, ",")
-}
-
-func (d *dockerDriver) addUserArgs(args []string, options *driver.RunOptions) []string {
-	if options.User != "" {
-		args = append(args, "-u", options.User)
-	}
-	return args
-}
-
-func (d *dockerDriver) addEnvArgs(args []string, options *driver.RunOptions) []string {
-	for k, v := range options.Env {
-		args = append(args, "-e", k+"="+v)
-	}
-	return args
-}
-
-func (d *dockerDriver) addInitArgs(args []string, options *driver.RunOptions) []string {
-	if options.Init != nil && *options.Init {
-		args = append(args, "--init")
-	}
-	return args
-}
-
-func (d *dockerDriver) addPrivilegedArgs(args []string, options *driver.RunOptions) []string {
-	if options.Privileged != nil && *options.Privileged {
-		args = append(args, "--privileged")
-	}
-	return args
-}
-
-func (d *dockerDriver) addCapabilityArgs(args []string, options *driver.RunOptions) []string {
-	for _, capAdd := range options.CapAdd {
-		args = append(args, "--cap-add", capAdd)
-	}
-	for _, securityOpt := range options.SecurityOpt {
-		args = append(args, "--security-opt", securityOpt)
-	}
-	return args
-}
-
-func (d *dockerDriver) addMountArgs(args []string, options *driver.RunOptions) ([]string, error) {
-	for _, mount := range options.Mounts {
-		if mount.Type == "bind" && mount.Source != "" {
-			if _, err := os.Stat(mount.Source); os.IsNotExist(err) {
-				return nil, fmt.Errorf("bind mount source path does not exist %s", mount.Source)
-			}
-		}
-		args = append(args, "--mount", mount.String())
-	}
-	return args, nil
-}
-
-func (d *dockerDriver) addIDEMountArgs(
-	args []string,
-	ide string,
-	ideOptions map[string]config2.OptionValue,
-) []string {
-	switch ide {
-	case string(config2.IDEGoland):
-		args = append(args, "--mount", jetbrains.NewGolandServer("", ideOptions).GetVolume())
-	case string(config2.IDERustRover):
-		args = append(
-			args,
-			"--mount",
-			jetbrains.NewRustRoverServer("", ideOptions).GetVolume(),
-		)
-	case string(config2.IDEPyCharm):
-		args = append(
-			args,
-			"--mount",
-			jetbrains.NewPyCharmServer("", ideOptions).GetVolume(),
-		)
-	case string(config2.IDEPhpStorm):
-		args = append(args, "--mount", jetbrains.NewPhpStorm("", ideOptions).GetVolume())
-	case string(config2.IDEIntellij):
-		args = append(args, "--mount", jetbrains.NewIntellij("", ideOptions).GetVolume())
-	case string(config2.IDECLion):
-		args = append(args, "--mount", jetbrains.NewCLionServer("", ideOptions).GetVolume())
-	case string(config2.IDERider):
-		args = append(args, "--mount", jetbrains.NewRiderServer("", ideOptions).GetVolume())
-	case string(config2.IDERubyMine):
-		args = append(
-			args,
-			"--mount",
-			jetbrains.NewRubyMineServer("", ideOptions).GetVolume(),
-		)
-	case string(config2.IDEWebStorm):
-		args = append(
-			args,
-			"--mount",
-			jetbrains.NewWebStormServer("", ideOptions).GetVolume(),
-		)
-	case string(config2.IDEDataSpell):
-		args = append(
-			args,
-			"--mount",
-			jetbrains.NewDataSpellServer("", ideOptions).GetVolume(),
-		)
-	}
-	return args
-}
-
-func (d *dockerDriver) addLabelArgs(
-	args []string,
-	workspaceId string,
-	options *driver.RunOptions,
-) []string {
-	labels := append(config.GetIDLabels(workspaceId, d.IDLabels), options.Labels...)
-	for _, label := range labels {
-		args = append(args, "-l", label)
-	}
-	return args
-}
-
-func (d *dockerDriver) addEntrypointArgs(args []string, options *driver.RunOptions) []string {
-	if options.Entrypoint != "" {
-		args = append(args, "--entrypoint", options.Entrypoint)
-	}
-	return args
-}
-
-func (d *dockerDriver) startContainer(
-	ctx context.Context,
-	dir string,
-	args []string,
-	writer io.Writer,
-) error {
-	log.Infof(
-		"running docker command: command=%s, args=%s, cwd=%s",
-		d.Docker.DockerCommand,
-		strings.Join(args, " "),
-		dir,
-	)
-
-	logHostEnvOnce(ctx, d.Docker)
-	logBindSources(args)
-
-	err := d.Docker.RunWithDir(ctx, dir, args, nil, writer, writer)
-	if err != nil {
-		log.Errorf(
-			"docker container failed to start: error=%v, command=%s, args=%s, cwd=%s",
-			err,
-			d.Docker.DockerCommand,
-			strings.Join(args, " "),
-			dir,
-		)
-		return fmt.Errorf("failed to start dev container: %w", err)
-	}
-	return nil
-}
-
-func resolveGPUAvailability(override string, d *dockerDriver) bool {
-	switch override {
-	case "true":
-		return true
-	case "false":
-		return false
-	default:
-		available, _ := d.Docker.GPUSupportEnabled()
-		return available
-	}
-}
-
-func appendGPUOptions(
-	parsedConfig *config.DevContainerConfig,
-	d *dockerDriver,
-	args []string,
-	gpuAvailabilityOverride string,
-) []string {
-	if parsedConfig.HostRequirements != nil {
-		gpuAvailable := resolveGPUAvailability(gpuAvailabilityOverride, d)
-		enableGPU, warnIfMissing := parsedConfig.HostRequirements.ShouldEnableGPU(gpuAvailable)
-		if enableGPU {
-			args = append(args, "--gpus", "all")
-		}
-		if warnIfMissing {
-			log.Warn("GPU required but not available on host")
-		}
-	}
-	return args
-}
-
-func (d *dockerDriver) getPodmanArgs(
-	options *driver.RunOptions,
-	parsedConfig *config.DevContainerConfig,
-) ([]string, error) {
-	if !d.Docker.GetRuntime().NeedsUserNamespaceArgs() {
-		return []string{}, nil
-	}
-
-	var args []string
-	args = d.addUsernsArgs(args, options)
-	args = d.addIdMappingArgs(args, options)
-	args = d.addKeepIdArgs(args, options, parsedConfig)
-	return args, nil
-}
-
-func (d *dockerDriver) addUsernsArgs(args []string, options *driver.RunOptions) []string {
-	if options.Userns != "" {
-		args = append(args, "--userns", options.Userns)
-	}
-	return args
-}
-
-func (d *dockerDriver) addIdMappingArgs(args []string, options *driver.RunOptions) []string {
-	for _, uidMap := range options.UidMap {
-		args = append(args, "--uidmap", uidMap)
-	}
-	for _, gidMap := range options.GidMap {
-		args = append(args, "--gidmap", gidMap)
-	}
-	return args
-}
-
-func (d *dockerDriver) addKeepIdArgs(
-	args []string,
-	options *driver.RunOptions,
-	parsedConfig *config.DevContainerConfig,
-) []string {
-	if d.hasIdMapping(options, parsedConfig) || options.Userns != "" {
-		return args
-	}
-
-	remoteUser := d.getRemoteUser(options, parsedConfig)
-	if remoteUser != "root" && remoteUser != "0" && os.Getuid() != 0 {
-		args = append(args, "--userns=keep-id")
-	}
-	return args
-}
-
-func (d *dockerDriver) hasIdMapping(
-	options *driver.RunOptions,
-	parsedConfig *config.DevContainerConfig,
-) bool {
-	if len(options.UidMap) > 0 || len(options.GidMap) > 0 {
-		return true
-	}
-
-	if parsedConfig != nil {
-		for _, arg := range parsedConfig.RunArgs {
-			if strings.Contains(arg, "--uidmap") || strings.Contains(arg, "--gidmap") {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-func (d *dockerDriver) getRemoteUser(
-	options *driver.RunOptions,
-	parsedConfig *config.DevContainerConfig,
-) string {
-	if parsedConfig != nil {
-		if parsedConfig.RemoteUser != "" {
-			return parsedConfig.RemoteUser
-		}
-		if parsedConfig.ContainerUser != "" {
-			return parsedConfig.ContainerUser
-		}
-	}
-	if options.User != "" {
-		return options.User
-	}
-	return "root"
-}
-
-func (d *dockerDriver) shouldUpdateUserUID(parsedConfig *config.DevContainerConfig) bool {
-	isLinux := runtime.GOOS == "linux"
-	hasUser := parsedConfig.ContainerUser != "" || parsedConfig.RemoteUser != ""
-
-	var shouldUpdate bool
-	switch {
-	case parsedConfig.UpdateRemoteUserUID != nil:
-		shouldUpdate = *parsedConfig.UpdateRemoteUserUID
-	case d.UpdateRemoteUserUIDDefault == "off":
-		shouldUpdate = false
-	default:
-		shouldUpdate = true
-	}
-
-	return isLinux && hasUser && shouldUpdate
-}
-
-func (d *dockerDriver) getContainerUser(parsedConfig *config.DevContainerConfig) string {
-	if parsedConfig.RemoteUser != "" {
-		return parsedConfig.RemoteUser
-	}
-	return parsedConfig.ContainerUser
-}
-
-func (d *dockerDriver) copyFileFromContainer(
-	ctx context.Context,
-	containerID, srcPath, dstPath string,
-	writer io.Writer,
-) error {
-	args := []string{"cp", fmt.Sprintf("%s:%s", containerID, srcPath), dstPath}
-	log.Debugf(
-		"copying file from container: command=%s, args=%s",
-		d.Docker.DockerCommand,
-		strings.Join(args, " "),
-	)
-	return d.Docker.Run(ctx, args, nil, writer, writer)
-}
-
-func (d *dockerDriver) copyFileToContainer(
-	ctx context.Context,
-	srcPath, containerID, dstPath string,
-	writer io.Writer,
-) error {
-	args := []string{"cp", srcPath, fmt.Sprintf("%s:%s", containerID, dstPath)}
-	log.Debugf(
-		"copying file to container: command=%s, args=%s",
-		d.Docker.DockerCommand,
-		strings.Join(args, " "),
-	)
-	return d.Docker.Run(ctx, args, nil, writer, writer)
-}
-
-type lineProcessor func(line string, fields []string) (modifiedLine string, shouldWrite bool, err error)
-
-func (d *dockerDriver) processColonDelimitedFile(
-	in *os.File,
-	out *os.File,
-	fieldCount int,
-	processor lineProcessor,
-) error {
-	scanner := bufio.NewScanner(in)
-
-	for scanner.Scan() {
-		line := scanner.Text()
-		fields := strings.SplitN(line, ":", fieldCount)
-
-		if len(fields) < fieldCount {
-			if _, err := fmt.Fprintf(out, "%s\n", line); err != nil {
-				return err
-			}
-			continue
-		}
-
-		modifiedLine, shouldWrite, err := processor(line, fields)
-		if err != nil {
-			return err
-		}
-
-		if shouldWrite {
-			if _, err := fmt.Fprintf(out, "%s\n", modifiedLine); err != nil {
-				return err
-			}
-		} else {
-			if _, err := fmt.Fprintf(out, "%s\n", line); err != nil {
-				return err
-			}
-		}
-	}
-
-	return scanner.Err()
-}
-
-type passwordFileUpdateParams struct {
-	passwdIn      *os.File
-	passwdOut     *os.File
-	containerUser string
-	localUid      string
-	localGid      string
-}
-
-// updatePasswdFile processes /etc/passwd, replacing the target user's UID/GID with local values.
-// It reads each line from passwdIn, and for lines matching containerUser, extracts the original
-// UID, GID, and home directory, then writes a modified entry with localUid and localGid to passwdOut.
-// All other lines are copied unchanged. Returns userInfo with the original container values, or an
-// error if the user is not found in the passwd file.
-func (d *dockerDriver) updatePasswdFile(params *passwordFileUpdateParams) (*user.User, error) {
-	info := &user.User{}
-
-	// parse passwd format: username:password:uid:gid:gecos:home:shell
-	processor := func(line string, fields []string) (string, bool, error) {
-		if fields[0] != params.containerUser {
-			return "", false, nil
-		}
-
-		info.Uid = fields[2]
-		info.Gid = fields[3]
-		info.HomeDir = fields[5]
-
-		modifiedLine := strings.Join([]string{
-			fields[0],
-			fields[1],
-			params.localUid,
-			params.localGid,
-			fields[4],
-			fields[5],
-			fields[6],
-		}, ":")
-		return modifiedLine, true, nil
-	}
-
-	if err := d.processColonDelimitedFile(
-		params.passwdIn,
-		params.passwdOut,
-		7,
-		processor,
-	); err != nil {
-		return nil, err
-	}
-
-	if info.Uid == "" {
-		return nil, fmt.Errorf("user %q not found in passwd", params.containerUser)
-	}
-
-	return info, nil
-}
-
-// updateGroupFile processes /etc/group, replacing entries with the target GID to use localGid.
-// It reads each line from groupIn, and for lines where the GID field matches containerGid,
-// writes a modified entry with localGid to groupOut. All other lines are copied unchanged.
-// Returns an error if scanning fails.
-func (d *dockerDriver) updateGroupFile(
-	groupIn *os.File,
-	groupOut *os.File,
-	containerGid, localGid string,
-) error {
-	// parse group format: groupname:password:gid:user_list
-	processor := func(line string, fields []string) (string, bool, error) {
-		if fields[2] != containerGid {
-			return "", false, nil
-		}
-
-		modifiedLine := strings.Join([]string{fields[0], fields[1], localGid, fields[3]}, ":")
-		return modifiedLine, true, nil
-	}
-
-	return d.processColonDelimitedFile(groupIn, groupOut, 4, processor)
-}
-
-func (d *dockerDriver) applyPermissions(
-	ctx context.Context,
-	containerID, localUid, localGid, containerHome string,
-	writer io.Writer,
-) error {
-	args := []string{"exec", "-u", "root", containerID, "chmod", "644", "/etc/passwd", "/etc/group"}
-	log.Debugf(
-		"modifying permissions of /etc/passwd and /etc/group: command=%s, args=%s",
-		d.Docker.DockerCommand,
-		strings.Join(args, " "),
-	)
-	if err := d.Docker.Run(ctx, args, nil, writer, writer); err != nil {
-		return err
-	}
-
-	if containerHome == "" {
-		log.Warnf(
-			"container home directory not found, skipping chown: containerID=%s",
-			containerID,
-		)
-		return nil
-	}
-
-	args = []string{
-		"exec",
-		"-u",
-		"root",
-		containerID,
-		"chown",
-		"-R",
-		fmt.Sprintf("%s:%s", localUid, localGid),
-		containerHome,
-	}
-	log.Debugf(
-		"running docker chown command: command=%s, args=%s",
-		d.Docker.DockerCommand,
-		strings.Join(args, " "),
-	)
-	return d.Docker.Run(ctx, args, nil, writer, writer)
-}
-
-type tempFiles struct {
-	passwdIn  *os.File
-	groupIn   *os.File
-	passwdOut *os.File
-	groupOut  *os.File
-}
-
-func (t *tempFiles) cleanup() {
-	if t.passwdIn != nil {
-		_ = t.passwdIn.Close()
-		_ = os.Remove(t.passwdIn.Name())
-	}
-	if t.groupIn != nil {
-		_ = t.groupIn.Close()
-		_ = os.Remove(t.groupIn.Name())
-	}
-	if t.passwdOut != nil {
-		_ = t.passwdOut.Close()
-		_ = os.Remove(t.passwdOut.Name())
-	}
-	if t.groupOut != nil {
-		_ = t.groupOut.Close()
-		_ = os.Remove(t.groupOut.Name())
-	}
-}
-
-func (d *dockerDriver) createTempFiles() (*tempFiles, error) {
-	files := &tempFiles{}
-	var err error
-
-	files.passwdIn, err = os.CreateTemp("", config2.BinaryName+"_container_passwd_in")
-	if err != nil {
-		return nil, err
-	}
-
-	files.groupIn, err = os.CreateTemp("", config2.BinaryName+"_container_group_in")
-	if err != nil {
-		files.cleanup()
-		return nil, err
-	}
-
-	files.passwdOut, err = os.CreateTemp("", config2.BinaryName+"_container_passwd_out")
-	if err != nil {
-		files.cleanup()
-		return nil, err
-	}
-
-	files.groupOut, err = os.CreateTemp("", config2.BinaryName+"_container_group_out")
-	if err != nil {
-		files.cleanup()
-		return nil, err
-	}
-
-	return files, nil
-}
-
-func (d *dockerDriver) fetchContainerFiles(
-	ctx context.Context,
-	containerID string,
-	files *tempFiles,
-	writer io.Writer,
-) error {
-	if err := d.copyFileFromContainer(
-		ctx,
-		containerID,
-		"/etc/passwd",
-		files.passwdIn.Name(),
-		writer,
-	); err != nil {
-		return err
-	}
-	return d.copyFileFromContainer(ctx, containerID, "/etc/group", files.groupIn.Name(), writer)
-}
-
-func (d *dockerDriver) processUserFiles(
-	files *tempFiles,
-	containerUser, localUid, localGid string,
-) (*user.User, error) {
-	passwdIn, err := os.Open(files.passwdIn.Name())
-	if err != nil {
-		return nil, err
-	}
-	defer func() { _ = passwdIn.Close() }()
-
-	info, err := d.updatePasswdFile(&passwordFileUpdateParams{
-		passwdIn:      passwdIn,
-		passwdOut:     files.passwdOut,
-		containerUser: containerUser,
-		localUid:      localUid,
-		localGid:      localGid,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	groupIn, err := os.Open(files.groupIn.Name())
-	if err != nil {
-		return nil, err
-	}
-	defer func() { _ = groupIn.Close() }()
-
-	return info, d.updateGroupFile(groupIn, files.groupOut, info.Gid, localGid)
-}
-
-func (d *dockerDriver) copyFilesToContainer(
-	ctx context.Context, containerID string, files *tempFiles, writer io.Writer,
-) error {
-	if err := d.copyFileToContainer(
-		ctx,
-		files.passwdOut.Name(),
-		containerID,
-		"/etc/passwd",
-		writer,
-	); err != nil {
-		return err
-	}
-	return d.copyFileToContainer(ctx, files.groupOut.Name(), containerID, "/etc/group", writer)
 }
